clean up bert
