+++ examples/bert/qnn/google_bert_qnn_fp32.json
@@ -24,21 +24,6 @@
     },
     "data_configs": [
         {
-            "name": "calibration_data",
-            "type": "HuggingfaceContainer",
-            "user_script": "google_bert_qnn.py",
-            "script_dir": ".",
-            "load_dataset_config": { "data_name": "glue", "subset": "mrpc", "split": "train[:10]" },
-            "pre_process_data_config": {
-                "type": "tokenize_dataset",
-                "input_cols": [ "sentence1", "sentence2" ],
-                "label_col": "label",
-                "seq_length": 128,
-                "model_name": "google-bert/bert-base-multilingual-cased"
-            },
-            "dataloader_config": { "batch_size": 1 }
-        },
-        {
             "name": "wiki_data",
             "type": "HuggingfaceContainer",
             "load_dataset_config": {
@@ -66,14 +51,19 @@
         "glue_evaluator": {
             "metrics": [
                 {
+                    "name": "accuracy_cpu",
+                    "type": "accuracy",
+                    "backend": "huggingface_metrics",
+                    "data_config": "wiki_data",
+                    "sub_types": [ { "name": "accuracy", "priority": 1 }, { "name": "f1" } ],
+                    "user_config": { "inference_settings": { "execution_provider": "CPUExecutionProvider" } }
+                },
+                {
                     "name": "accuracy_qnn",
                     "type": "accuracy",
                     "backend": "huggingface_metrics",
                     "data_config": "wiki_data",
-                    "sub_types": [
-                        { "name": "accuracy", "priority": 1, "goal": { "type": "max-degradation", "value": 0.05 } },
-                        { "name": "f1" }
-                    ],
+                    "sub_types": [ { "name": "accuracy", "priority": 2 }, { "name": "f1" } ],
                     "user_config": {
                         "inference_settings": {
                             "onnx": {
@@ -83,7 +73,7 @@
                                         "backend_path": "QnnHtp.dll",
                                         "htp_performance_mode": "burst",
                                         "htp_graph_finalization_optimization_mode": "3",
-                                        "offload_graph_io_quantization": "0"
+                                        "enable_htp_fp16_precision": "1"
                                     }
                                 ]
                             }
@@ -91,10 +81,18 @@
                     }
                 },
                 {
+                    "name": "latency_cpu",
+                    "type": "latency",
+                    "sub_types": [
+                        { "name": "avg", "priority": 3, "metric_config": { "warmup_num": 20, "repeat_test_num": 100 } }
+                    ],
+                    "user_config": { "inference_settings": { "execution_provider": "CPUExecutionProvider" } }
+                },
+                {
                     "name": "latency_qnn",
                     "type": "latency",
                     "sub_types": [
-                        { "name": "avg", "priority": 2, "metric_config": { "warmup_num": 20, "repeat_test_num": 100 } }
+                        { "name": "avg", "priority": 4, "metric_config": { "warmup_num": 20, "repeat_test_num": 100 } }
                     ],
                     "user_config": {
                         "inference_settings": {
@@ -105,7 +103,7 @@
                                         "backend_path": "QnnHtp.dll",
                                         "htp_performance_mode": "burst",
                                         "htp_graph_finalization_optimization_mode": "3",
-                                        "offload_graph_io_quantization": "0"
+                                        "enable_htp_fp16_precision": "1"
                                     }
                                 ]
                             }
@@ -116,26 +114,10 @@
         }
     },
     "passes": {
-        "conversion": { "type": "OnnxConversion", "target_opset": 20, "dynamic": true, "use_dynamo_exporter": false },
+        "conversion": { "type": "OnnxConversion", "target_opset": 20 },
         "surgery": {
             "type": "GraphSurgeries",
             "surgeries": [ { "surgeon": "ReplaceAttentionMaskValue" }, { "surgeon": "MatMulAddToGemm" } ]
-        },
-        "quantization": {
-            "type": "OnnxStaticQuantization",
-            "data_config": "calibration_data",
-            "quant_preprocess": true,
-            "calibrate_method": "MinMax",
-            "activation_type": "uint16",
-            "precision": "uint8"
-        },
-        "cb": {
-            "type": "EPContextBinaryGenerator",
-            "provider_options": {
-                "htp_graph_finalization_optimization_mode": "3",
-                "offload_graph_io_quantization": "0"
-            },
-            "weight_sharing": true
         }
     },
     "host": "host_system",
@@ -145,5 +127,5 @@
     "clean_evaluation_cache": true,
     "evaluator": "glue_evaluator",
     "evaluate_input_model": false,
-    "output_dir": "models/google_bert_base_multilingual_cased_qdq_ctx"
+    "output_dir": "models/google_bert_base_multilingual_cased_fp32"
 }
