+++ examples/bert/qnn/google_bert_qnn_fp32_ctx.json
@@ -55,10 +55,7 @@
                     "type": "accuracy",
                     "backend": "huggingface_metrics",
                     "data_config": "wiki_data",
-                    "sub_types": [
-                        { "name": "accuracy", "priority": 1, "goal": { "type": "max-degradation", "value": 0.05 } },
-                        { "name": "f1" }
-                    ],
+                    "sub_types": [ { "name": "accuracy", "priority": 1 }, { "name": "f1" } ],
                     "user_config": {
                         "inference_settings": {
                             "onnx": {
@@ -101,15 +98,14 @@
         }
     },
     "passes": {
-        "conversion": { "type": "OnnxConversion", "target_opset": 20, "dynamic": true, "use_dynamo_exporter": false },
+        "conversion": { "type": "OnnxConversion", "target_opset": 20 },
         "surgery": {
             "type": "GraphSurgeries",
             "surgeries": [ { "surgeon": "ReplaceAttentionMaskValue" }, { "surgeon": "MatMulAddToGemm" } ]
         },
         "cb": {
             "type": "EPContextBinaryGenerator",
-            "provider_options": { "htp_graph_finalization_optimization_mode": "3", "enable_htp_fp16_precision": "1" },
-            "weight_sharing": true
+            "provider_options": { "htp_graph_finalization_optimization_mode": "3", "enable_htp_fp16_precision": "1" }
         }
     },
     "host": "host_system",
