+++ examples/bert/qnn/google_bert_qnn_qdq.json
@@ -24,6 +24,21 @@
     },
     "data_configs": [
         {
+            "name": "calibration_data",
+            "type": "HuggingfaceContainer",
+            "user_script": "google_bert_qnn.py",
+            "script_dir": ".",
+            "load_dataset_config": { "data_name": "glue", "subset": "mrpc", "split": "train[:10]" },
+            "pre_process_data_config": {
+                "type": "tokenize_dataset",
+                "input_cols": [ "sentence1", "sentence2" ],
+                "label_col": "label",
+                "seq_length": 128,
+                "model_name": "google-bert/bert-base-multilingual-cased"
+            },
+            "dataloader_config": { "batch_size": 1 }
+        },
+        {
             "name": "wiki_data",
             "type": "HuggingfaceContainer",
             "load_dataset_config": {
@@ -55,10 +70,7 @@
                     "type": "accuracy",
                     "backend": "huggingface_metrics",
                     "data_config": "wiki_data",
-                    "sub_types": [
-                        { "name": "accuracy", "priority": 1, "goal": { "type": "max-degradation", "value": 0.05 } },
-                        { "name": "f1" }
-                    ],
+                    "sub_types": [ { "name": "accuracy", "priority": 1 }, { "name": "f1" } ],
                     "user_config": { "inference_settings": { "execution_provider": "CPUExecutionProvider" } }
                 },
                 {
@@ -66,10 +78,7 @@
                     "type": "accuracy",
                     "backend": "huggingface_metrics",
                     "data_config": "wiki_data",
-                    "sub_types": [
-                        { "name": "accuracy", "priority": 2, "goal": { "type": "max-degradation", "value": 0.05 } },
-                        { "name": "f1" }
-                    ],
+                    "sub_types": [ { "name": "accuracy", "priority": 2 }, { "name": "f1" } ],
                     "user_config": {
                         "inference_settings": {
                             "onnx": {
@@ -79,7 +88,7 @@
                                         "backend_path": "QnnHtp.dll",
                                         "htp_performance_mode": "burst",
                                         "htp_graph_finalization_optimization_mode": "3",
-                                        "enable_htp_fp16_precision": "1"
+                                        "offload_graph_io_quantization": "0"
                                     }
                                 ]
                             }
@@ -109,7 +118,7 @@
                                         "backend_path": "QnnHtp.dll",
                                         "htp_performance_mode": "burst",
                                         "htp_graph_finalization_optimization_mode": "3",
-                                        "enable_htp_fp16_precision": "1"
+                                        "offload_graph_io_quantization": "0"
                                     }
                                 ]
                             }
@@ -120,10 +129,18 @@
         }
     },
     "passes": {
-        "conversion": { "type": "OnnxConversion", "target_opset": 20, "dynamic": true, "use_dynamo_exporter": false },
+        "conversion": { "type": "OnnxConversion", "target_opset": 20 },
         "surgery": {
             "type": "GraphSurgeries",
             "surgeries": [ { "surgeon": "ReplaceAttentionMaskValue" }, { "surgeon": "MatMulAddToGemm" } ]
+        },
+        "quantization": {
+            "type": "OnnxStaticQuantization",
+            "data_config": "calibration_data",
+            "quant_preprocess": true,
+            "calibrate_method": "MinMax",
+            "activation_type": "uint16",
+            "precision": "uint8"
         }
     },
     "host": "host_system",
@@ -133,5 +150,5 @@
     "clean_evaluation_cache": true,
     "evaluator": "glue_evaluator",
     "evaluate_input_model": false,
-    "output_dir": "models/google_bert_base_multilingual_cased_fp32"
+    "output_dir": "models/google_bert_base_multilingual_cased_qdq"
 }
