+++ examples/bert/qnn/google_bert_qnn_qdq_ctx.json
@@ -66,25 +66,11 @@
         "glue_evaluator": {
             "metrics": [
                 {
-                    "name": "accuracy_cpu",
-                    "type": "accuracy",
-                    "backend": "huggingface_metrics",
-                    "data_config": "wiki_data",
-                    "sub_types": [
-                        { "name": "accuracy", "priority": 1, "goal": { "type": "max-degradation", "value": 0.05 } },
-                        { "name": "f1" }
-                    ],
-                    "user_config": { "inference_settings": { "execution_provider": "CPUExecutionProvider" } }
-                },
-                {
                     "name": "accuracy_qnn",
                     "type": "accuracy",
                     "backend": "huggingface_metrics",
                     "data_config": "wiki_data",
-                    "sub_types": [
-                        { "name": "accuracy", "priority": 2, "goal": { "type": "max-degradation", "value": 0.05 } },
-                        { "name": "f1" }
-                    ],
+                    "sub_types": [ { "name": "accuracy", "priority": 1 }, { "name": "f1" } ],
                     "user_config": {
                         "inference_settings": {
                             "onnx": {
@@ -102,18 +88,10 @@
                     }
                 },
                 {
-                    "name": "latency_cpu",
-                    "type": "latency",
-                    "sub_types": [
-                        { "name": "avg", "priority": 3, "metric_config": { "warmup_num": 20, "repeat_test_num": 100 } }
-                    ],
-                    "user_config": { "inference_settings": { "execution_provider": "CPUExecutionProvider" } }
-                },
-                {
                     "name": "latency_qnn",
                     "type": "latency",
                     "sub_types": [
-                        { "name": "avg", "priority": 4, "metric_config": { "warmup_num": 20, "repeat_test_num": 100 } }
+                        { "name": "avg", "priority": 2, "metric_config": { "warmup_num": 20, "repeat_test_num": 100 } }
                     ],
                     "user_config": {
                         "inference_settings": {
@@ -135,7 +113,7 @@
         }
     },
     "passes": {
-        "conversion": { "type": "OnnxConversion", "target_opset": 20, "dynamic": true, "use_dynamo_exporter": false },
+        "conversion": { "type": "OnnxConversion", "target_opset": 20 },
         "surgery": {
             "type": "GraphSurgeries",
             "surgeries": [ { "surgeon": "ReplaceAttentionMaskValue" }, { "surgeon": "MatMulAddToGemm" } ]
@@ -147,6 +125,13 @@
             "calibrate_method": "MinMax",
             "activation_type": "uint16",
             "precision": "uint8"
+        },
+        "cb": {
+            "type": "EPContextBinaryGenerator",
+            "provider_options": {
+                "htp_graph_finalization_optimization_mode": "3",
+                "offload_graph_io_quantization": "0"
+            }
         }
     },
     "host": "host_system",
@@ -156,5 +141,5 @@
     "clean_evaluation_cache": true,
     "evaluator": "glue_evaluator",
     "evaluate_input_model": false,
-    "output_dir": "models/google_bert_base_multilingual_cased_qdq"
+    "output_dir": "models/google_bert_base_multilingual_cased_qdq_ctx"
 }
